Text.Regex.PCRE.ByteString

Description

This exports instances of the high level API and the medium level API of compile,execute, and regexec.

Types

data Regex

A compiled regular expression

type MatchOffset = Int

0 based index from start of source, or (-1) for unused

type MatchLength = Int

non-negative length of a match

newtype CompOption

Constructors

CompOption CInt

newtype ExecOption

Constructors

ExecOption CInt

data ReturnCode

type WrapError = (ReturnCode, String)

Miscellaneous

unusedOffset :: MatchOffset

getVersion :: Maybe String

Version string of PCRE library

NOTE: The Maybe type is used for historic reasons; practically, getVersion is never Nothing.

Medium level API functions

compile

Arguments

:: CompOption
(summed together)

-> ExecOption
(summed together)

-> ByteString
The regular expression to compile

-> IO (Either (MatchOffset, String) Regex)
Returns: the compiled regular expression

Compiles a regular expression

execute

Arguments

:: Regex
Compiled regular expression

-> ByteString
String to match against

-> IO (Either WrapError (Maybe (Array Int (MatchOffset, MatchLength))))
Returns: Nothing if the regex did not match the string, or: Just an array of (offset,length) pairs where index 0 is whole match, and the rest are the captured subexpressions.

Matches a regular expression against a buffer, returning the buffer indicies of the match, and any submatches

| Matches a regular expression against a string

regexec

Arguments

:: Regex
Compiled regular expression

-> ByteString
String to match against

-> IO (Either WrapError (Maybe (ByteString, ByteString, ByteString, [ByteString])))

CompOption flags

compBlank :: CompOption

compAnchored :: CompOption

compAutoCallout :: CompOption

compCaseless :: CompOption

compDollarEndOnly :: CompOption

compDotAll :: CompOption

compExtended :: CompOption

compExtra :: CompOption

compFirstLine :: CompOption

compMultiline :: CompOption

compNoAutoCapture :: CompOption

compUngreedy :: CompOption

compUTF8 :: CompOption

compNoUTF8Check :: CompOption

ExecOption flags

execBlank :: ExecOption

execAnchored :: ExecOption

execNotBOL :: ExecOption

execNotEOL :: ExecOption

execNotEmpty :: ExecOption

execNoUTF8Check :: ExecOption

execPartial :: ExecOption

Orphan instances

RegexLike Regex ByteString

Instance details

Methods

matchOnce :: Regex -> ByteString -> Maybe MatchArray

matchAll :: Regex -> ByteString -> [MatchArray]

matchCount :: Regex -> ByteString -> Int

matchTest :: Regex -> ByteString -> Bool

matchAllText :: Regex -> ByteString -> [MatchText ByteString]

matchOnceText :: Regex -> ByteString -> Maybe (ByteString, MatchText ByteString, ByteString)

RegexContext Regex ByteString ByteString

Instance details

Methods

match :: Regex -> ByteString -> ByteString

matchM :: MonadFail m => Regex -> ByteString -> m ByteString

RegexMaker Regex CompOption ExecOption ByteString

Instance details

Methods

makeRegex :: ByteString -> Regex

makeRegexOpts :: CompOption -> ExecOption -> ByteString -> Regex

makeRegexM :: MonadFail m => ByteString -> m Regex

makeRegexOptsM :: MonadFail m => CompOption -> ExecOption -> ByteString -> m Regex